requirements.txt
django
djangorestframework
pillow
psycopg2-binary
python-decouple
dj_rest_auth
drf-yasg
django-filter
djangorestframework-jwt
uuid
djangorestframework-simplejwt==5.2.2
django-star-ratings
celery
